refactor(pbr): split background-material.ts and lazy-load solid skybox
Extract the solid-color skybox into background-solid-skybox.ts, dyn-imported
only when !skipSkybox so scenes like scene13 (skipSkybox:true) no longer pay
for the skybox shader module.

Delete monolithic background-material.ts, split into:
- skybox-geometry.ts: shared createSkyboxBuffers, buildSkyboxWorldMatrix, createBuf
- cubemap-skybox-material.ts: createCubemapSkyboxMaterial (DDS + HDR only)
- createSkyboxMaterial inlined into background-solid-skybox.ts (its only caller)

Ground-only scenes no longer load cubemap/solid skybox code; solid-skybox
scenes no longer load cubemap material code.

Results (raw KB):
- scene1:  92.6 -> 91.0 (-1.6)
- scene7:  101.0 -> 99.9 (-1.1, back under ceiling)
- scene13: 90.5 -> 86.8 (-3.7)
- scene14: 92.8 -> 91.2 (-1.6)
- scene20: 81.3 -> 79.7 (-1.6)
- scene23: 79.7 -> 78.1 (-1.6)

No API changes, no ceiling changes, no reference changes.
