+++ packages/babylon-lite/src/material/pbr/background-dds-skybox.ts
@@ -6,7 +6,8 @@ import type { EngineContextInternal } from "../../engine/engine.js";
 import type { Renderable } from "../../render/renderable.js";
 import { getOrCreateSampler } from "../../resource/gpu-pool.js";
 import { computeSkyboxGeometry } from "./background-renderable.js";
-import { createSkyboxBuffers, buildSkyboxWorldMatrix, createCubemapSkyboxMaterial } from "./background-material.js";
+import { createSkyboxBuffers, buildSkyboxWorldMatrix } from "./skybox-geometry.js";
+import { createCubemapSkyboxMaterial } from "./cubemap-skybox-material.js";
 import { WGSL_SCENE_UNIFORMS_PBR, WGSL_DITHER } from "../../shader/wgsl-helpers.js";
 import ddsSkyboxVertSrc from "../../../shaders/skybox-dds.vertex.wgsl?raw";
 import ddsSkyboxFragSrc from "../../../shaders/skybox-dds.fragment.wgsl?raw";
