+++ packages/babylon-lite/src/material/pbr/background-hdr-skybox.ts
@@ -7,7 +7,8 @@ import type { SceneContext } from "../../scene/scene.js";
 import type { EngineContextInternal } from "../../engine/engine.js";
 import type { EnvironmentTextures } from "../../loader-env/load-env.js";
 import type { Renderable } from "../../render/renderable.js";
-import { createSkyboxBuffers, buildSkyboxWorldMatrix, createCubemapSkyboxMaterial } from "./background-material.js";
+import { createSkyboxBuffers, buildSkyboxWorldMatrix } from "./skybox-geometry.js";
+import { createCubemapSkyboxMaterial } from "./cubemap-skybox-material.js";
 import { computeSkyboxGeometry } from "./background-renderable.js";
 import skyboxVertSrc from "../../../shaders/skybox.vertex.wgsl?raw";
 import skyboxHdrFragSrc from "../../../shaders/skybox-hdr.fragment.wgsl?raw";
